[refactor] : token에서 멤버 id 추출 및 자잘한 수정

+++ src/main/java/backend/univfit/domain/apply/application/HomeAnnouncementServiceImpl.java
@@ -22,7 +22,7 @@ public class HomeAnnouncementServiceImpl implements HomeAnnouncementService {
                 .stream()
                 .map(objects -> {
                     AnnouncementEntity announcement = (AnnouncementEntity) objects[0];
-                    Integer likeCount = (Integer) objects[1];
+                    Long likeCount = (Long) objects[1];
                     return PopularAnnouncementResponse.of(
                             announcement.getId(),
                             announcement.getScholarShipImage(),
